Spur_Gear_Sizing_streamlit/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
plotly==5.20.0
streamlit==1.32.2
streamlit-player==0.1.5
cadquery==2.4.0
